AG-10672 Fixes
